Initial upload of Perl scripts used in Gray et al. (2012).
great_overlaps.pl compares a BED file to a regulatory domain file generated by the GREAT createRegulatoryDomains tool, and counts the number of times the center of the provided BED regions overlap the regulatory domains. The result can be used with the GREAT calculateBinomialP tool to determine if the number of overlaps is significant.

sam_sorted_to_fragment_wig.pl converts a sorted, paired-end SAM read alignment file to a fragment overlap WIG file.

wig_peak_summits.pl will find the highest-scored section of a WIG file for each region provided in a BED file. For each region, it will report the score, the start, end, distance of the center from the 5' end of the region, and a relative distance where 0 is the start position and 1 is the end position of the input BED region.

wig_pileup_bed.pl will generate pileups of WIG data relative to the 5' end of a list of regions in BED format (that is, it takes the strand column into account to determine the direction of the pileup). The use inputs the distance up and downstream from the 5' end that should be calculated. The output is a simple tab-separated file with the relative position and cumulative score on each line.